fix: simplify asset plugin
