+++ packages/builder/builder/src/plugins/index.ts
@@ -1,3 +1,9 @@
+import {
+  MEDIA_EXTENSIONS,
+  FONT_EXTENSIONS,
+  IMAGE_EXTENSIONS,
+} from '@modern-js/builder-shared';
+
 export const plugins = {
   cleanOutput: () =>
     import('./cleanOutput').then(m => m.builderPluginCleanOutput()),
@@ -13,7 +19,14 @@ export const plugins = {
     import('./splitChunks').then(m => m.builderPluginSplitChunks()),
   bundleAnalyzer: () =>
     import('./bundleAnalyzer').then(m => m.builderPluginBundleAnalyzer()),
-  font: () => import('./font').then(m => m.builderPluginFont()),
-  image: () => import('./image').then(m => m.builderPluginImage()),
-  media: () => import('./media').then(m => m.builderPluginMedia()),
+  font: () =>
+    import('./asset').then(m => m.builderAssetPlugin('font', FONT_EXTENSIONS)),
+  image: () =>
+    import('./asset').then(m =>
+      m.builderAssetPlugin('image', IMAGE_EXTENSIONS),
+    ),
+  media: () =>
+    import('./asset').then(m =>
+      m.builderAssetPlugin('media', MEDIA_EXTENSIONS),
+    ),
 };
